[ACTP] move network path bundle to subdirectory
